Переход от int к long числам, вывод зон, новые 4 типа, для ровного деления начало и конца поля, перенумерование типов

+++ PRNcompression/Services/Interfaces/IPRNService.cs
@@ -11,10 +11,10 @@ namespace PRNcompression.Services.Interfaces
     }
     internal interface IPRNService
     {
-        int PRNGeneration(byte type, int size);
+        long PRNGeneration(byte type, int size);
         int GetNumberType(int num, int dimensionality);
         int GetNumberLength(int num);
         byte[] FieldCharacterization(byte dimensionality);
-        CompressedInfo Compression(int number, int numberLength, ref List<bool> list);
+        CompressedInfo Compression(long number, int numberLength, ref List<bool> list);
     }
 }
